🚚 move: firebase auth 파일 분리
